Renamed tchar not to conflict with system ones
