+++ src/burner/libretro/fba_tchar.h
